# docs/swagger.yaml
basePath: /api/v1
definitions:
  subscriptions_rest_internal_domain.Date:
    description: The date
    properties:
      month:
        type: integer
      year:
        type: integer
    type: object
  subscriptions_rest_internal_domain.PriceResponse:
    description: The price
    properties:
      price:
        type: integer
    type: object
  subscriptions_rest_internal_domain.Subscribtion:
    description: The subscribtion
    properties:
      end_date:
        $ref: '#/definitions/subscriptions_rest_internal_domain.Date'
      price:
        type: integer
      service_name:
        type: string
      start_date:
        $ref: '#/definitions/subscriptions_rest_internal_domain.Date'
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a service for checking subscriptions.
  license:
    name: Subscriptions Service 1.0
  title: Subscriptions Service
  version: "1.0"
paths:
  /subscribtions:
    delete:
      description: Удаляет подписку
      parameters:
      - description: deleteSubscribtion
        in: body
        name: Entry
        required: true
        schema:
          $ref: '#/definitions/subscriptions_rest_internal_domain.Subscribtion'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Delete subscription
    get:
      description: Получает подписку
      parameters:
      - description: getSubscribtion
        in: body
        name: Entry
        required: true
        schema:
          $ref: '#/definitions/subscriptions_rest_internal_domain.Subscribtion'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscriptions_rest_internal_domain.Subscribtion'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Get subscription
    post:
      description: Создает подписку
      parameters:
      - description: createSubscribtion
        in: body
        name: Entry
        required: true
        schema:
          $ref: '#/definitions/subscriptions_rest_internal_domain.Subscribtion'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Create subscription
    put:
      description: Обновляет подписку
      parameters:
      - description: updateSubscribtion
        in: body
        name: Entry
        required: true
        schema:
          $ref: '#/definitions/subscriptions_rest_internal_domain.Subscribtion'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Update subscription
  /subscribtions/all:
    get:
      description: Получает все подписки
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/subscriptions_rest_internal_domain.Subscribtion'
            type: array
        "500":
          description: Internal Server Error
      summary: Get all subscriptions
  /subscribtions/price:
    get:
      description: Получает стоимость подписок за указанный период
      parameters:
      - description: getSubscribtionsPrice
        in: body
        name: Entry
        required: true
        schema:
          $ref: '#/definitions/subscriptions_rest_internal_domain.Subscribtion'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscriptions_rest_internal_domain.PriceResponse'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Get subscriptions price
swagger: "2.0"